--- secure-api-gateway-ob-uk-rs-resource-store/secure-api-gateway-ob-uk-rs-resource-store-repo/src/main/java/com/forgerock/sapi/gateway/rs/resource/store/repo/mongo/v4/accounts/directdebits/FRDirectDebitRepositoryCustom.java
@@ -1,37 +1,0 @@
-/*
- * Copyright © 2020-2024 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.sapi.gateway.rs.resource.store.repo.mongo.v4.accounts.directdebits;
-
-import com.forgerock.sapi.gateway.ob.uk.common.datamodel.account.FRExternalPermissionsCode;
-import com.forgerock.sapi.gateway.rs.resource.store.repo.entity.v4.account.FRDirectDebit;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.repository.query.Param;
-
-import java.util.List;
-
-public interface FRDirectDebitRepositoryCustom {
-
-    Page<FRDirectDebit> byAccountIdWithPermissions(
-            @Param("accountId") String accountId,
-            @Param("permissions") List<FRExternalPermissionsCode> permissions,
-            Pageable pageable);
-
-    Page<FRDirectDebit> byAccountIdInWithPermissions(
-            @Param("accountIds") List<String> accountIds,
-            @Param("permissions") List<FRExternalPermissionsCode> permissions,
-            Pageable pageable);
-}
