--- secure-api-gateway-ob-uk-rs-resource-store/secure-api-gateway-ob-uk-rs-resource-store-repo/src/main/java/com/forgerock/sapi/gateway/rs/resource/store/repo/mongo/v4/accounts/party/FRPartyRepositoryImpl.java
@@ -1,56 +1,0 @@
-/*
- * Copyright © 2020-2024 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.sapi.gateway.rs.resource.store.repo.mongo.v4.accounts.party;
-
-import com.forgerock.sapi.gateway.ob.uk.common.datamodel.account.FRExternalPermissionsCode;
-import com.forgerock.sapi.gateway.rs.resource.store.repo.entity.v4.account.FRParty;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository("V4.0.0FRPartyRepositoryImpl")
-public class FRPartyRepositoryImpl implements FRPartyRepositoryCustom {
-    @Autowired
-    @Lazy
-    private FRPartyRepository party1Repository;
-
-    @Override
-    public FRParty byAccountIdWithPermissions(String accountId, List<FRExternalPermissionsCode> permissions) {
-        return filter(party1Repository.findByAccountId(accountId), permissions);
-    }
-
-    @Override
-    public FRParty byUserIdWithPermissions(String userId, List<FRExternalPermissionsCode> permissions) {
-        return filter(party1Repository.findByUserId(userId), permissions);
-    }
-
-    @Override
-    public Page<FRParty> byAccountIdInWithPermissions(List<String> accountIds, List<FRExternalPermissionsCode> permissions, Pageable pageable) {
-        return filter(party1Repository.findByAccountIdIn(accountIds, pageable), permissions);
-    }
-
-    private Page<FRParty> filter(Page<FRParty> parties, List<FRExternalPermissionsCode> permissions) {
-        return parties;
-    }
-
-    private FRParty filter(FRParty party, List<FRExternalPermissionsCode> permissions) {
-        return party;
-    }
-}
