--- secure-api-gateway-ob-uk-rs-resource-store/secure-api-gateway-ob-uk-rs-resource-store-repo/src/main/java/com/forgerock/sapi/gateway/rs/resource/store/repo/mongo/v4/accounts/standingorders/FRStandingOrderRepository.java
@@ -1,41 +1,0 @@
-/*
- * Copyright © 2020-2024 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.sapi.gateway.rs.resource.store.repo.mongo.v4.accounts.standingorders;
-
-import com.forgerock.sapi.gateway.rs.resource.store.repo.entity.v4.account.FRStandingOrder;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Component;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Set;
-
-@Component("V4.0.0FRStandingOrderRepository")
-public interface FRStandingOrderRepository extends MongoRepository<FRStandingOrder, String>, FRStandingOrderRepositoryCustom {
-
-    Page<FRStandingOrder> findByAccountId(@Param("accountId") String accountId, Pageable pageable);
-
-    Page<FRStandingOrder> findByAccountIdIn(@Param("accountIds") List<String> accountIds, Pageable pageable);
-
-    Long deleteStandingOrderByAccountId(@Param("accountId") String accountId);
-
-    List<FRStandingOrder> findByStatusIn(@Param("statuses") Collection<FRStandingOrder.StandingOrderStatus> statuses);
-
-    Long countByAccountIdIn(Set<String> accountIds);
-}
