+++ secure-api-gateway-ob-uk-rs-server/src/main/java/com/forgerock/sapi/gateway/ob/uk/rs/server/api/obie/account/v4_0_0/directdebits/DirectDebitsApiController.java
@@ -23,8 +23,8 @@ import com.forgerock.sapi.gateway.ob.uk.rs.server.v4.common.util.AccountDataInte
 import com.forgerock.sapi.gateway.ob.uk.rs.server.v4.common.util.PaginationUtil;
 import com.forgerock.sapi.gateway.ob.uk.rs.server.service.account.consent.AccountResourceAccessService;
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.account.v3_1_10.AccountAccessConsent;
-import com.forgerock.sapi.gateway.rs.resource.store.repo.entity.v4.account.FRDirectDebit;
-import com.forgerock.sapi.gateway.rs.resource.store.repo.mongo.v4.accounts.directdebits.FRDirectDebitRepository;
+import com.forgerock.sapi.gateway.rs.resource.store.repo.entity.account.FRDirectDebit;
+import com.forgerock.sapi.gateway.rs.resource.store.repo.mongo.accounts.directdebits.FRDirectDebitRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
